pyproject.toml
[project]
name = "TransformersProject"
version = "0.1.0"
description = "Homework 4 - Transformers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "matplotlib",
    "scipy",
    "numba",
    "tqdm",
    "seaborn",
    "requests",
    "scikit-learn",
    "torch",
    "torchvision",
    "transformers[torch]==4.57.3",
    "pytorch-triton-rocm",
    "evaluate",
    "huggingface",
    "pyarrow",
    "fastparquet",
    "python-dotenv",
]

# [[tool.uv.index]]
# name = "pytorch-cu128"
# url = "https://download.pytorch.org/whl/cu128"
# explicit = true

# [tool.uv.sources]
# torch = [
#   { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
# ]
# torchvision = [
#   { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
# ]

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]
pytorch-triton-rocm = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]